+++ RTRendererCUDAQT/RTRendererCUDAQT/Texture.h
@@ -6,7 +6,7 @@
 class Texture
 {
 public:
-	__device__ virtual Vec3 value(double u, double v, const Vec3& p) const = 0;
+	__device__ virtual Vec3 value(float u, float v, const Vec3& p) const = 0;
 };
 
 class ConstantTexture : public Texture
@@ -16,8 +16,8 @@ public:
 
 	__device__ ConstantTexture() {}
 	__device__ ConstantTexture(Vec3 c) : color(c) {}
-	__device__ ConstantTexture(double r, double g, double b) : color(Vec3(r, g ,b)) {}
-	__device__ virtual Vec3 value(double u, double v, const Vec3& p) const
+	__device__ ConstantTexture(float r, float g, float b) : color(Vec3(r, g ,b)) {}
+	__device__ virtual Vec3 value(float u, float v, const Vec3& p) const
 	{
 		return color;
 	}
@@ -32,9 +32,9 @@ public:
 	__device__ CheckerTexture() :even(new ConstantTexture(1, 1, 1)), odd(new ConstantTexture(0, 0, 0)) {}
 	__device__ CheckerTexture(Texture* t0, Texture* t1) :even(t0), odd(t1) {}
 
-	__device__ virtual Vec3 value(double u, double v, const Vec3& p) const
+	__device__ virtual Vec3 value(float u, float v, const Vec3& p) const
 	{
-		double sines = sin(10 * p.e[0]) * sin(10 * p.e[1]) * sin(10 * p.e[2]);
+		float sines = sin(10 * p.e[0]) * sin(10 * p.e[1]) * sin(10 * p.e[2]);
 		if (sines < 0)
 			return odd->value(u, v, p);
 		else
@@ -46,11 +46,11 @@ class NoiseTexture : public Texture
 {
 public:
 	Perlin noise;
-	double scale;
+	float scale;
 
 	__device__ NoiseTexture() :scale(1) {}
-	__device__ NoiseTexture(double sc) :scale(sc) {}
-	__device__ virtual Vec3 value(double u, double v, const Vec3& p) const
+	__device__ NoiseTexture(float sc) :scale(sc) {}
+	__device__ virtual Vec3 value(float u, float v, const Vec3& p) const
 	{
 		// return Vec3(1, 1, 1) * noise.turbulence(p * 4);
 		return Vec3(1, 1, 1) * 0.5 * (1 + sin(scale * p.e[2] + 10 * noise.turbulence(p)));
@@ -66,7 +66,7 @@ public:
 	__device__ ImageTexture(unsigned char* d, int x, int y)
 		:data(d), imgX(x), imgY(y) {}
 
-	__device__ virtual Vec3 value(double u, double v, const Vec3& p) const
+	__device__ virtual Vec3 value(float u, float v, const Vec3& p) const
 	{
 		int i = u * imgX;
 		int j = (1 - v) * imgY;
