+++ RTRendererCUDAQT/RTRendererCUDAQT/Volumes.h
@@ -7,19 +7,19 @@ class ConstantMedium : public Hittable
 {
 public:
 	Hittable* boundary;
-	double density;
+	float density;
 	Material* phaseFunc;
 
-	__device__ ConstantMedium(Hittable* b, double d, Texture* a) : boundary(b), density(d), phaseFunc(new Isotropic(a))	{}
+	__device__ ConstantMedium(Hittable* b, float d, Texture* a) : boundary(b), density(d), phaseFunc(new Isotropic(a))	{}
 
-	__device__ virtual bool hit(const Ray& r, double tMin, double tMax, HitRecord& rec, curandState* localRandState) const;
+	__device__ virtual bool hit(const Ray& r, float tMin, float tMax, HitRecord& rec, curandState* localRandState) const;
 	__device__ virtual bool boundingBox(AABB& box) const
 	{
 		return boundary->boundingBox(box);
 	}
 };
 
-__device__ bool ConstantMedium::hit(const Ray& r, double tMin, double tMax, HitRecord& rec, curandState* localRandState) const
+__device__ bool ConstantMedium::hit(const Ray& r, float tMin, float tMax, HitRecord& rec, curandState* localRandState) const
 {
 	HitRecord rec1, rec2;
 	if (boundary->hit(r, -DBL_MAX, DBL_MAX, rec1, localRandState))
@@ -31,8 +31,8 @@ __device__ bool ConstantMedium::hit(const Ray& r, double tMin, double tMax, HitR
 			if (rec1.t >= rec2.t) return false;
 			rec1.t = rec1.t < 0 ? 0 : rec1.t;
 
-			double distance = (rec2.t - rec1.t) * r.direction.length();
-			double hitAt = - (1/density) * log(curand_uniform(localRandState));
+			float distance = (rec2.t - rec1.t) * r.direction.length();
+			float hitAt = - (1/density) * log(curand_uniform(localRandState));
 
 			if (hitAt < distance)
 			{
